+++ torchensemble/_constants.py
@@ -173,6 +173,21 @@ __predict_doc = """
 """
 
 
+__vectorize_doc = """
+    Return the vectorization result of the ensemble using functorch. Details
+    available at `functorch model ensembling <https://pytorch.org/functorch/stable/notebooks/ensembling.html>`_.
+
+    Returns
+    -------
+    fmodel : FunctionalModuleWithBuffers
+        Functional version of one of the models in the ensemble.
+    params : tuple
+        Tuple of stacked model parameters in the ensemble.
+    buffers : tuple
+        Tuple of buffers, empty if not exists.
+"""  # noqa: E501
+
+
 __classification_forward_doc = """
     Parameters
     ----------